Add faster calc of number of coins
Before, the coins of a certain stack where added one by one. Before each
addition it was checked if the adding of the coin face value to the
stack value was above the target value. If not, the stack size was
increased by one.

Now, the total number of coins of a particular face value (stack) are
calculated in one step, by division and dropping of the decimal places.

This is way more straight forward and should reduce the number of needed
operations a lot.