Change CoinStack from inheritance to composition
Before, the CoinStack class was derived from the Coin class. While this
worked, it did not make sense. The CoinStack is not a type of Coin.

Now, the CoinStack is a base class derived from `object`. The relation
to the type of coin that is collected in the stack is defined in the
field `coin_type`. This makes way more sense. I still show that
there is a connection between CoinStack and Coin and that there can only
be one type of coin in a coin stack. But, I am avoiding the confusing
and pretty much wrong pattern of inheriting from a class without have a
"IS-A" relationship.